// src/app/core/services/contact/types.ts
export interface EmailForm {
    recruiter: string;
    name: string;
    email: string;
    ocupation: string;
    message: string;
    celphone: string;
    companyName: string;
    companyWebsite: string;
    buttonLabel: string;
}

export interface ButtonLabel {
    sendEmail: string;
    emailSent: string;
}